+++ app/src/main/java/com/do55anto5/moviestreaming/di/RepositoryModule.kt
@@ -1,6 +1,6 @@
 package com.do55anto5.moviestreaming.di
 
-import com.do55anto5.moviestreaming.data.remote.SignupRepositoryImpl
+import com.do55anto5.moviestreaming.data.remote.repository.authentication.SignupRepositoryImpl
 import com.do55anto5.moviestreaming.domain.remote.repository.authentication.SignupRepository
 import org.koin.dsl.module
